feat: responsive redesign of Students
Split 747-line file into Students (246), StudentFormModal (329),
DeleteStudentModal (54), StudentTable (101). Remove palette object,
use UI components, add responsive table columns hidden on mobile.
